requirements.txt
requests~=2.32.4
python-dotenv~=1.1.0
logging~=0.4.9.6
colorlog~=6.9.0